fix: fix failing storybook tests
